requirements.txt
streamlit
pandas
pystrich
Pillow